feat(geometry): the B-rep's construction math is double too — primitives, parameters, crossings
The stage the previous commit named. Widening storage alone bought nothing end to end,
because a double vertex holding a float-accurate point is still float-accurate. This moves
the arithmetic that PLACES geometry.

- All 14 primitive builders place their vertices in double: makeBox, makeOpenBox,
  makeCylinder, makeCone, makeSphere, makeFacetedCylinder, makeFacetedSphere, makeTube,
  makePyramid, extrudeProfile, loftProfiles, twistExtrude, revolveProfile and its partial
  form. They had to move TOGETHER -- the previous commit measured what happens otherwise:
  widening makeCylinder alone improved the cylinder (5.9e-8 -> 4.4e-8) and made the box
  WORSE (1.24e-7), because a double-built cylinder's seam vertices agree with a float-built
  box's less well than two float-built ones did.
- Curve/Surface eval take a DOUBLE parameter, and radii are double. A double position
  evaluated at a float angle is only accurate to float, which is what the first cut
  measured: 6e-8 relative, unmoved, because the angle was the limit and not the point.
- Edge::t0/t1, splitEdge's fraction, setEdgeArc's radius, paramOnCurve's return, the
  circle/segment crossing solver and every angle constant are double. Found by chasing an
  imprinted vertex sitting at z = 1.0000000596046448 -- which is 1 + 2^-24, the float just
  above 1.0, so a float intermediate was still placing it.
- Serialization v4 carries radii and edge parameters as double alongside the positions, and
  still reads v1-v3 as float.

MEASURED: the worst mismatch between an edge's curve and the vertices it must reproduce, on
a cylinder-through-box imprint, is now 4.37e-8 relative and IDENTICAL for both operands at
every scale (1, 100, 10000) -- where it was 1.24e-7 for the box against 4.4e-8 for the
cylinder while the two were built differently. Both sides agreeing exactly is the signal
that they are now built the same way.

WHAT DID NOT HAPPEN, and it corrects this arc's own plan. The previous commit predicted that
once the primitives moved, Tolerance's defaults -- sized to float at kDefaultAbsolute=1e-5,
kDefaultRelative=1e-6 -- could finally be re-tightened. Measured: they cannot. Tightening to
1e-6/1e-7 fails 6 tests; to 1e-7/1e-8, 23. So something still needs the loose band, and it is
not the B-rep's construction error, which is now 4.4e-8 with over twenty times that headroom.
Reverted, and left as the open question rather than forced.

Also unresolved: that residual 4.37e-8 is itself float-flavoured -- one diagnostic showed an
arc parameter holding exactly float pi (3.1415927410125732), and 0.5*sin(float pi) is
4.37e-8 exactly. So at least one float source remains in the chain and is NOT yet located.
The improvement is real and measured; the ceiling is not yet fully lifted.

Test: ctest --test-dir build -> 2495/2495 (5 Vulkan-HW skipped).

+++ apps/Nexus-Modeling/src/kernel/include/nexus/geometry/AnalyticBRep.h
@@ -229,7 +229,7 @@ public:
     // endpoint must recognise that and reuse the endpoint vertex itself (see
     // imprintCurve) rather than rely on this floor, which would otherwise
     // manufacture a new vertex measurably offset from the true crossing.
-    uint32_t splitEdge(uint32_t edgeId, float t, Tolerance tol = {});
+    uint32_t splitEdge(uint32_t edgeId, double t, Tolerance tol = {});
 
     // Euler operator — split a face by a new edge (a Line) between two
     // NON-ADJACENT vertices of its outer loop, producing two faces that inherit
@@ -287,7 +287,7 @@ public:
     // the curve still reproduces the endpoint vertices — so checkGeometry holds
     // and toMesh(subdivisions) now tessellates it smoothly. Returns false if the
     // endpoints are not on the circle within tolerance.
-    bool setEdgeArc(uint32_t edgeId, const Vec3& center, const Vec3& axis, float radius,
+    bool setEdgeArc(uint32_t edgeId, const Vec3& center, const Vec3& axis, double radius,
                     Tolerance tol = {});
 
     // Attach a parameter-space trim curve (pcurve) to a coedge: a straight
